requirements.txt
streamlit>=1.42.0
pandas>=2.1.4
numpy>=1.26.3
folium>=0.15.1
matplotlib>=3.8.2
geopandas>=0.14.2
fiona>=1.9.5
shapely>=2.0.2
streamlit-folium>=0.15.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.18.0
pydantic>=2.5.2 
python-dotenv>=1.0.1

